// src/helper/types/validators_registry/validator.d.ts
/* tslint:disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

export type HumanAddr = string;
export type Uint128 = string;

export interface Validator {
  active: boolean;
  address: HumanAddr;
  delegation_limit: Uint128;
  total_delegated?: Uint128 & string;
  [k: string]: unknown;
}