// src/helper/types/cw20_token/token_init_msg.d.ts
/* tslint:disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

export type HumanAddr = string;
export type Uint128 = string;

export interface TokenInitMsg {
  decimals: number;
  hub_contract: HumanAddr;
  initial_balances: Cw20CoinHuman[];
  mint?: MinterResponse | null;
  name: string;
  symbol: string;
  [k: string]: unknown;
}
export interface Cw20CoinHuman {
  address: HumanAddr;
  amount: Uint128;
  [k: string]: unknown;
}
export interface MinterResponse {
  /**
   * cap is how many more tokens can be issued by the minter
   */
  cap?: Uint128 | null;
  minter: HumanAddr;
  [k: string]: unknown;
}