// src/helper/types/validators_registry/query_msg.d.ts
/* tslint:disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

export type QueryMsg = {
  get_validators_for_delegation: {
    [k: string]: unknown;
  };
  [k: string]: unknown;
};