// src/helper/types/basset_reward/query_msg.d.ts
/* tslint:disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

export type QueryMsg =
  | {
      config: {
        [k: string]: unknown;
      };
      [k: string]: unknown;
    }
  | {
      state: {
        [k: string]: unknown;
      };
      [k: string]: unknown;
    }
  | {
      accrued_rewards: {
        address: HumanAddr;
        [k: string]: unknown;
      };
      [k: string]: unknown;
    }
  | {
      holder: {
        address: HumanAddr;
        [k: string]: unknown;
      };
      [k: string]: unknown;
    }
  | {
      holders: {
        limit?: number | null;
        start_after?: HumanAddr | null;
        [k: string]: unknown;
      };
      [k: string]: unknown;
    };
export type HumanAddr = string;